projeto_eva_completo_final/projeto_eva/requirements.txt
# Core ML/AI
torch>=2.0.0
transformers>=4.30.0
llama-cpp-python>=0.2.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Memory and Storage
chromadb>=0.4.0
numpy>=1.24.0
pandas>=2.0.0
sqlite3

# Voice Processing
openai-whisper>=20230314
piper-tts>=1.2.0
coqui-tts>=0.15.0
sounddevice>=0.4.6
pyaudio>=0.2.11

# UI and System Integration
gradio>=3.40.0
psutil>=5.9.0
pyautogui>=0.9.54
opencv-python>=4.8.0
pillow>=10.0.0
pynput>=1.7.6

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0
asyncio
aiofiles>=23.0.0
click>=8.0.0
colorama>=0.4.6

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: GPU monitoring
nvidia-ml-py3>=7.352.0
